Advance commit revision for couchdb, release 2.5.0
This advances the couchdb commit, for the release 2.5.0, to include a
single change:

https://github.com/couchbase/couchdb/commit/a425d9742ae28414d06ec03dac02b85d0bac41c3

This was experienced by a customer (CBSE-893). While it happened in an
evaluation/testing scenario with a not much likely to happen sequence of
steps in production environment, it's worth it as others might do the
same mistake in a an environment with more more writes to a particular
vbucket where they added by mistake a design document via the memcached
API (instead of REST API / UI).

Change-Id: I5393855ff4f26c5c67d2f973ea73eb80d4451445
Reviewed-on: http://review.couchbase.org/31772
Reviewed-by: Philip Labee <philip@couchbase.com>
Tested-by: Philip Labee <philip@couchbase.com>