MB-13094: Sherlock based manifests with 3.0.2 revisions
The `toy/toy-volker-302.xml` is a manifest that uses the current
Sherlock manifest as a basis, but uses the revisions of the Couchbase
core components of the 3.0.2 release.

This will be used as a baseline to find performance regressions in
Sherlock.

The `toy/toy-volker.xml` manifest is based on the `toy/toy-volker-302.xml`
manifest, but with the view-engine using the most recent Sherlock revisions
(couchdb, couchstore and geocouch).

Change-Id: I9066b4f23033c61d13c619a537241df6a84f918b
Reviewed-on: http://review.couchbase.org/46254
Reviewed-by: Volker Mische <volker.mische@gmail.com>
Tested-by: Volker Mische <volker.mische@gmail.com>