Freeze the SHA1 for memcached and platform
There is no outstanding bugs reported for these projects and
we've got patches we want to go into 4.5.1. It is easier for
us to deal with that now rather than have the patch waiting in
gerrit and merge them in a couple of weeks.

If it ever turns out that we need to build a patch for a
bug in these projects we should create a branch at that
time.

Change-Id: Icd690ae8ec8895204c7cb4ea7c1df377060ee14a
Reviewed-on: http://review.couchbase.org/64316
Reviewed-by: Chris Hillery <ceej@couchbase.com>
Tested-by: Chris Hillery <ceej@couchbase.com>