MB-20742 - rollback bleve to build 1055 / before MB-20589
Build 1056 had test blocker regression (MB-20742) and looks like it's
related to recent bleve component upgrade for feature MB-20589.

So, this change is a rollback to the bleve component where it was at
previous build 1055.

Not all the components of build 1056 were rolled back (blevex,
segment, phosphor), as they don't look currently like suspects for the
test regression.

Change-Id: I7351c166cccf1ff9abe244df49cb68bb1e29c5a1
Reviewed-on: http://review.couchbase.org/67189
Reviewed-by: Steve Yen <steve.yen@gmail.com>
Tested-by: Steve Yen <steve.yen@gmail.com>